dataframe_merge.py is now full self-sustained
